requirements.txt
Django==5.0.6
django-imagekit==5.0.0
djangorestframework==3.15.1
drf-yasg==1.21.7
pre-commit==3.7.1
python-decouple==3.8
